% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_analysis_preprocess.R
\name{calculate_cell_exp}
\alias{calculate_cell_exp}
\title{Calculate expected cell type-specific expression for one spot}
\usage{
calculate_cell_exp(spot_count, spot_nnls_output, spot_ref)
}
\arguments{
\item{spot_count}{A vector of gene expression level for one spot.}

\item{spot_nnls_output}{A vector of relative cell number for the spot gotten by deconvolution.}

\item{spot_ref}{A matrix of expression profile for cell types as reference.}
}
\description{
This function calculate expected cell type-specific expression based on deconvolution results of relative
cell number, cell type-specific expression profile and expression vector for one spot.
}
