% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_analysis_preprocess.R
\name{calculate_expected_celltype_exp}
\alias{calculate_expected_celltype_exp}
\title{Calculate expected cell type-specific expression for one spatial transcriptomics sample}
\usage{
calculate_expected_celltype_exp(
  filepath = getwd(),
  st_count,
  st_coords,
  ref_celltype_gep,
  loop = 2,
  alpha = 0.2,
  p_cut = 1e-05,
  max_cores = 8
)
}
\arguments{
\item{filepath}{A string of file path name which saves the outputs of STANE loop results.}

\item{st_count}{Expression matrix for single-cell level data.}

\item{st_coords}{The dataframe containing coordinate of spots.}

\item{ref_celltype_gep}{The cell type-specific expression profile for reference.}

\item{loop}{The number of loop of which deconvolution and niche effect results will be used for calculation.}

\item{alpha}{A weight that controls the influential level of niche effects on gene expression.}

\item{p_cut}{A threshold of p-values for niche effect estimation. Niche effects with regression p-values lower
than the p_cut is considered as significant.}

\item{max_cores}{Number of cores used for parallel computing. If max_cores < 1, stop using parallel computing.}
}
\description{
This function gives a list of expected expression matrix of all cell types for one spatial trancriptomics
sample, considering niche effects calculated in the STANE major process.
}
\examples{
cell_count_list<-calculate_expected_celltype_exp(filepath = getwd(),st_count = st_count, st_coords = st_coords, ref_celltype_gep = ref_celltype_count, loop = 2, max_cores = 6)
}
