% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_analysis_preprocess.R
\name{calculate_localmoranI_bivariat}
\alias{calculate_localmoranI_bivariat}
\title{Calculate co-upregulated bi-variant local Moran's I}
\usage{
calculate_localmoranI_bivariat(input_x, input_y, weight_mtx)
}
\arguments{
\item{input_x, input_y}{A vector of expression level for ligand/receptor genes.}

\item{weight_mtx}{A matrix of weights that controls influences of the distance between neighbor cell and target
cell.}
}
\description{
This function calculate the co-upregulated bi-variant local Moran's I for ligand and receptor gene expression
in the neighbor and target cell type. Only if the ligand and receptor genes are both higher than the average
expressio level, is the local Moran's I higher than 0.
}
