% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_deconvolution_with_nicheDE.R
\name{calculate_nnls_cut}
\alias{calculate_nnls_cut}
\title{Perform cell type-specific deconvlution to estimate relative cell numbers for spots}
\usage{
calculate_nnls_cut(
  spot_count,
  marker_ref_celltype_gep,
  cutoff = 0.1,
  cell_cut_list = NULL
)
}
\arguments{
\item{spot_count}{A vector of gene expression level for one spot.}

\item{marker_ref_celltype_gep}{The reference cell type-specific expression profile of marker genes.}

\item{cutoff}{A threshold for cell proportion; cell type with higher proportion than the cutoff
exists in the spot.}

\item{cell_cut_list}{A vector of cut-off values for each cell type. If NULL, will use the same
threshold defined by cutoff.}
}
\description{
This function use Non-negative least squares regression to estimate relative numbers of each
cell type in the spots based on cell type-specific expression profiles.
}
\examples{
# cell_cut_list=NULL
data(st_count)
nnls_output <- calculate_nnls_cut(spot_count = st_count[,1], marker_ref_celltype_gep = marker_ref_celltype_count)
# Set up cell_cut_list
cell_cut_list<-c(0.2,0.4,0.2,0.1,0.1,0.1)
names(cell_cut_list)<-colnames(marker_ref_celltype_count)
nnls_output <- calculate_nnls_cut(spot_count = st_count[,1], marker_ref_celltype_gep = marker_ref_celltype_count, cell_cut_list = cell_cut_list)
}
