% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_effect_estimation.R
\name{check_deconvolution_coverge}
\alias{check_deconvolution_coverge}
\title{Check whether the deconvlution results converge.}
\usage{
check_deconvolution_coverge(filepath = getwd(), loop)
}
\arguments{
\item{filepath}{A string of file path name to save the outputs of each loop.}

\item{loop}{The number of loop you want to check its converge level compared with previous loop.}
}
\description{
This function is used to check whether the deconvlution results from two adjcent loops converge
by calculating Root mean squared error.
}
