% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_deconvolution_with_nicheDE.R
\name{determine_cell_exist}
\alias{determine_cell_exist}
\title{Determine whether a cell type existing in one spot}
\usage{
determine_cell_exist(prop_vec, cutoff = 0.1, cell_cut_list = NULL)
}
\arguments{
\item{prop_vec}{A vector of cell proportion gotten by cell type-specific deconvolution.}

\item{cutoff}{A threshold for cell proportion; cell type with higher proportion than the cutoff
exists in the spot.}

\item{cell_cut_list}{A vector of cut-off values for each cell type. If NULL, will use the same
threshold defined by cutoff.}
}
\description{
This function is used to determine whether a cell type exists in the spot based on the estimated
cell proportion and cut-off values.
}
