% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_deconvolution_with_nicheDE.R
\name{generate_dist_mtx}
\alias{generate_dist_mtx}
\title{Generate distance matrix for spatial sequencing data}
\usage{
generate_dist_mtx(coord_df, use_rank = T)
}
\arguments{
\item{coord_df}{The dataframe containing coordinate of spots.}

\item{use_rank}{Logical. If true, the distance to a target spot is represented by the rank of distances for
all neighbor spots; else the distance is represented by Euclidean distance. Default is TRUE.}
}
\description{
This function generates distance matrix based on coordinates of spots from spatial transcriptomics
data.
}
\examples{
data(st_coords)
dist_mat<-generate_dist_mtx(coord_df = st_coords)
}
