% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_prepare.R
\name{generate_marker_celltype_gep}
\alias{generate_marker_celltype_gep}
\title{Generate reference expression profile for cell type's marker genes}
\usage{
generate_marker_celltype_gep(
  ref_celltype_gep,
  sc_count,
  label,
  max_p = 1e-30,
  top_count = 100
)
}
\arguments{
\item{ref_celltype_gep}{The reference expression profile which is generated by `generate_ref_celltype_gep()`.}

\item{sc_count}{Expression matrix for single-cell level data.}

\item{label}{A vector of cell labels for all cells in the sc_count.}

\item{max_p}{The threshold of p-value used for extracting significantly upregulated marker genes for cell types.}

\item{top_count}{Number of marker genes selected for each cell type.}
}
\description{
This function generates reference expression profile for marker genes significantly unregulated
in each cell type, which will be applied in the following cell-type deconvolution step.
}
\examples{
data(sc_count)
data(sc_label)
ref_celltype_count<-generate_ref_celltype_gep(sc_count = sc_count, label = sc_label)
marker_ref_celltype_count<-generate_marker_celltype_gep(ref_celltype_gep = ref_celltype_count, sc_count = sc_count, label = sc_label)
}
