% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_prepare.R
\name{generate_ref_celltype_gep}
\alias{generate_ref_celltype_gep}
\title{Generate reference expression profile from sc-level data}
\usage{
generate_ref_celltype_gep(sc_count, label, normalize = T)
}
\arguments{
\item{sc_count}{Expression matrix for single-cell level data.}

\item{label}{A vector of cell labels for all cells in the sc_count.}

\item{normalize}{Logical. Whether do normalization on sc_count. Default is TRUE.}
}
\description{
This function generates reference expression profile from single-cell level expression data,
which will be applied in the following niche-effect estimation steps.
}
\examples{
data(sc_count)
data(sc_label)
ref_celltype_count<-generate_ref_celltype_gep(sc_count = sc_count, label = sc_label)
}
