% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_deconvolution_with_nicheDE.R
\name{generate_spot_exp}
\alias{generate_spot_exp}
\title{Generate spot-specific cell expression profiles}
\usage{
generate_spot_exp(
  alpha = 0.2,
  spotID,
  sig_beta_list,
  marker_ref_celltype_gep,
  neighbor_prop
)
}
\arguments{
\item{alpha}{A constant that controls the influential level of niche effects on gene expression.}

\item{spotID}{The spotID of the index spot to calculate spot-specific expression profile.}

\item{sig_beta_list}{A dataframe containing information of niche effects gotten by previous loop.}

\item{marker_ref_celltype_gep}{The reference cell type-specific expression profile of marker genes
gotten from `generate_marker_celltype_gep()`.}

\item{neighbor_prop}{A matrix of neighbor cell abundance gotten by `neighbor_abundance()`.}
}
\description{
This function generates cell type-specific expression profile for one spot based on its neighbor
cell abundance, niche effects from surroundings and initial expression profile gotten by sc-level
data.
}
