% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_analysis_preprocess.R
\name{get_correlated_LR}
\alias{get_correlated_LR}
\title{Find the most significant ligand-receptor interaction on cell-cell niche effects.}
\usage{
get_correlated_LR(
  filepath = getwd(),
  st_coords,
  cell_count_list,
  LR_list,
  nicheDE_res,
  keep_moranI = T,
  target_celltype,
  neighbor_celltype,
  p_cut = 1e-05,
  cor_cut = 0.001,
  count_cut = 10,
  max_cores = 8
)
}
\arguments{
\item{filepath}{A string of file path name to save the outputs of this loop.}

\item{st_coords}{The dataframe containing coordinate of spots.}

\item{cell_count_list}{A list of expression profile for cell types calculated by function `calculate_expected_celltype_exp()`.}

\item{LR_list}{A dataframe containing lignd and receptor genes.}

\item{nicheDE_res}{A dataframe of niche effect estimation gotten by STANE process.}

\item{keep_moranI}{Logical. If true, save the bi-variant Moran's I results in the predefined file path. Default is TRUE.}

\item{target_celltype}{A cell type name which receives niche effects and acts as receptor.}

\item{neighbor_celltype}{A cell type name which gives niche effects and acts as ligand.}

\item{p_cut}{A threshold of p-values for niche effect estimation. Niche effects with regression p-values lower
than the p_cut is considered as significant.}

\item{cor_cut}{A threshold of correlation test p-values to find correlated genes with ligand-receptor activity.}

\item{count_cut}{A threshold of expression level for ligand and receptor genes in cell types. Only genes higher than this
cut-off will be used to do ligand-receptor analysis.}

\item{max_cores}{Number of cores used for parallel computing. If max_cores < 1, stop using parallel computing.}
}
\description{
This function use correlation analysis and enrichment test to find the most significant ligand-receptor interactions
which play important roles on gene expression alterations caused by niche effects.
}
\examples{
data(LR_list)
loop = 2
cell_count_list<-calculate_expected_celltype_exp(filepath = getwd(),st_count = st_count, st_coords = st_coords, ref_celltype_gep = ref_celltype_count, loop = loop, max_cores = 6)
nicheDE_result<-read.csv(paste0(filepath,"/nicheDE_result/nicheDE_res_loop_",loop,".csv"))
cor_LR_moranI_res<-get_correlated_LR(st_coords = st_coords,cell_count_list = cell_count_list,LR_list = LR_list,nicheDE_res = nicheDE_result,target_celltype = "Tumor",neighbor_celltype = "Myeloid",max_cores = 6)
}
