% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_deconvolution_with_nicheDE.R
\name{neighbor_abundance}
\alias{neighbor_abundance}
\title{Calculate abundance of neighbor cell types for spots}
\usage{
neighbor_abundance(feature_mat, dist_mat, spotID, c = 0.5, consider_self = T)
}
\arguments{
\item{feature_mat}{A matrix of cell type-specific proportions for spots.}

\item{dist_mat}{A matrix of spot-spot distance generated by `generate_dist_mtx()`.}

\item{spotID}{The spotID for the index spot from feature_mat and dist_mat.}

\item{c}{A constant that controls the weight of spot-spot distance during abundance calculation.}

\item{consider_self}{Logical. Whether considering cells in the index spot itself while calculating
cell abundance for this spot.}
}
\description{
This function calculate abundance of neighbor cell types for each spot based on cell proportions
and distance of neighbor spots.
}
\examples{
neighbor_abundance(feature_mat = prop_estimate, dist_mat = dist_mat, spotID = row.names(prop_estimate)[1], c = 0.05)

}
