% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_effect_estimation.R
\name{run_lasso_reg}
\alias{run_lasso_reg}
\title{Estimate niche effects on one gene's expression}
\usage{
run_lasso_reg(
  select_gene,
  count,
  ref_celltype_gep,
  neighbor_abundance_mat,
  cell_deconv_res
)
}
\arguments{
\item{select_gene}{One of gene name from expression matrix selected for niche effect estimation.}

\item{count}{A vector of gene expression level for the selected gene.}

\item{ref_celltype_gep}{The cell type-specific expression profile for reference.}

\item{neighbor_abundance_mat}{A matrix of neighbor cell abundance.}

\item{cell_deconv_res}{A matrix of relative cell type-specific number calculated by cell deconvolution.}
}
\description{
This function uses cell type-specific expression profile and deconvolution results to estimate the
level of niche effects on one gene's expression. In this process, a modified lasso regression, the
Induced Smoothed Lasso model, is used to evaluate altered expression level caused by neighbor cells,
and a cross-validation method is applied to select lambda for regression.
}
\examples{
data(st_count)
select_gene<-row.names(st_count)[1]
res<-run_lasso_reg(select_gene=select_gene,count = st_count,ref_celltype_gep = ref_celltype_count,neighbor_abundance_mat = mat_neighbor_abundance,cell_deconv_res = nnls_output)
res
}
\references{
Cilluffo G, Sottile G, Muggeo V (2019). _The R package islasso: estimation and
hypothesis testing in lasso regression_, volume 0 number 0. doi:10.13140/RG.2.2.16360.11521 <https://doi.org/10.13140/RG.2.2.16360.11521>.

Friedman J, Tibshirani R, Hastie T (2010). “Regularization Paths for Generalized Linear Models
via Coordinate Descent.” _Journal of Statistical Software_, *33*(1), 1-22. doi:10.18637/jss.v033.i01 <https://doi.org/10.18637/jss.v033.i01>.
}
