% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_STANE.R
\name{run_STANE}
\alias{run_STANE}
\title{Run the major STANE process for one loop}
\usage{
run_STANE(
  filepath = getwd(),
  st_count,
  st_coords,
  ref_celltype_gep,
  marker_ref_celltype_gep,
  cell_cut_list = NULL,
  loop = 1,
  c = 0.05,
  consider_self = T,
  alpha = 0.2,
  p_cut = 1e-05,
  prop_cut = 0.1,
  max_cores = 8
)
}
\arguments{
\item{filepath}{A string of file path name to save the outputs of this loop.}

\item{st_count}{Expression matrix for single-cell level data.}

\item{st_coords}{The dataframe containing coordinate of spots.}

\item{ref_celltype_gep}{The cell type-specific expression profile for reference.}

\item{marker_ref_celltype_gep}{The cell type-specific expression profile of marker genes for reference.}

\item{cell_cut_list}{A vector of cut-off values for each cell type. If NULL, will use the same
threshold defined by cutoff.}

\item{loop}{The number of loop for iteration. If loop = 1, start the STANE process without previous niche-effect
information; else if loop > 1, consider previous niche effects in the process.}

\item{c}{A weight that controls the weight of spot-spot distance during abundance calculation.}

\item{consider_self}{Logical. Whether considering cells in the index spot itself while calculating
cell abundance for this spot.}

\item{alpha}{A weight that controls the influential level of niche effects on gene expression.}

\item{p_cut}{A threshold of p-values for niche effect estimation. Niche effects with regression p-values lower
than the p_cut is considered as significant.}

\item{prop_cut}{A threshold for cell proportion deconvolution; cell type with higher proportion than the cutoff
exists in the spot.}

\item{max_cores}{Number of cores used for parallel computing. If max_cores < 1, stop using parallel computing.}
}
\description{
This function runs the major process of one loop for STANE, including cell deconvolution on spot-level
sptial transcriptomics data, niche effect estimation caused by cell-cell interactions and spot-specific cell
expression profile adjusted by niche effects. The results of this loop will be saved in the defined file path.
}
\examples{
data(sc_label)
data(sc_count)
ref_celltype_count<-generate_ref_celltype_gep(sc_count = sc_count, label = sc_label)
marker_ref_celltype_count<-generate_marker_celltype_gep(ref_celltype_gep = ref_celltype_count, sc_count = sc_count, label = sc_label)
data(st_count)
data(st_coords)
run_STANE(filepath=getwd(),st_count = st_count,st_coords = st_coords,ref_celltype_gep = ref_celltype_count,marker_ref_celltype_gep = marker_ref_celltype_count,loop = 1,max_cores=5)
}
